using_tensorflow/requirements.txt
absl-py==0.6.1
astor==0.7.1
Click==7.0
cloudpickle==0.6.1
cycler==0.10.0
dask==0.20.1
decorator==4.3.0
Flask==1.0.2
gast==0.2.0
grpcio==1.16.0
gunicorn==19.9.0
h5py==2.8.0
imageio==2.4.1
imgaug==0.2.6
itsdangerous==1.1.0
Jinja2==2.10
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
networkx==2.2
nltk==3.3
numpy==1.15.4
opencv-python==3.4.3.18
Pillow==5.3.0
protobuf==3.6.1
pyparsing==2.3.0
python-dateutil==2.7.5
PyWavelets==1.0.1
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
toolz==0.9.0
Werkzeug==0.14.1
